Ajustes de dtos e extensions
