+++ src/main/kotlin/br/com/fiap/postech/fastfood/adapter/inbound/CheckoutController.kt
@@ -1,7 +1,9 @@
 package br.com.fiap.postech.fastfood.adapter.inbound
 
 import br.com.fiap.postech.fastfood.adapter.inbound.extension.toCheckoutDTO
+import br.com.fiap.postech.fastfood.adapter.inbound.extension.toCheckoutResponse
 import br.com.fiap.postech.fastfood.adapter.inbound.request.CheckoutRequest
+import br.com.fiap.postech.fastfood.adapter.inbound.response.CheckoutResponse
 import br.com.fiap.postech.fastfood.application.domain.dtos.CheckoutDTO
 import br.com.fiap.postech.fastfood.application.ports.interfaces.CheckoutServicePort
 import org.springframework.http.HttpStatus
@@ -15,7 +17,7 @@ class CheckoutController (
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    fun enviaParaFila(@RequestBody checkoutRequest: CheckoutRequest): CheckoutDTO {
-        return checkoutServicePort.enviaParaFila(checkoutRequest.toCheckoutDTO())
+    fun enviaParaFila(@RequestBody checkoutRequest: CheckoutRequest): CheckoutResponse {
+        return checkoutServicePort.enviaParaFila(checkoutRequest.toCheckoutDTO()).toCheckoutResponse()
     }
 }
