+++ src/main/kotlin/br/com/fiap/postech/fastfood/adapter/outbound/infrastructure/database/repositories/impl/CheckoutRepositoryImpl.kt
@@ -1,10 +1,9 @@
 package br.com.fiap.postech.fastfood.adapter.outbound.infrastructure.database.repositories.impl
 
+import br.com.fiap.postech.fastfood.adapter.inbound.extension.toCheckoutEntity
+import br.com.fiap.postech.fastfood.adapter.inbound.extension.toCheckoutModel
 import br.com.fiap.postech.fastfood.adapter.outbound.infrastructure.database.entities.CheckoutEntity
-import br.com.fiap.postech.fastfood.adapter.outbound.infrastructure.database.entities.PedidoEntity
 import br.com.fiap.postech.fastfood.adapter.outbound.infrastructure.database.repositories.CheckoutRepositorySpring
-import br.com.fiap.postech.fastfood.application.domain.extension.toCheckoutEntity
-import br.com.fiap.postech.fastfood.application.domain.extension.toCheckoutModel
 import br.com.fiap.postech.fastfood.application.domain.extension.toPedidoEntity
 import br.com.fiap.postech.fastfood.application.domain.models.Checkout
 import br.com.fiap.postech.fastfood.application.domain.models.Pedido
